refactor: abstracts MenuList to Parent and Group components
